requirements.txt
# -------- CORE EXTRACTORS --------
PyMuPDF==1.24.8
pdfminer.six==20231228
pdfplumber==0.11.1
python-docx==0.8.11
Pillow==10.2.0
pytesseract==0.3.10

# -------- OCR/IMAGE IMPROVEMENTS --------
opencv-python==4.10.0.84
scikit-image==0.22.0

# -------- PARSING, CLEANING & NLP --------
spacy==3.7.2
pyspellchecker==0.8.1
regex==2024.5.15

# -------- EVALUATION + UTILITIES --------
numpy==1.26.4
pandas==2.2.3
tqdm==4.66.4
rapidfuzz==3.9.4

# -------- GENERAL UTILITIES --------
python-dotenv==1.0.1
loguru==0.7.2
rich==13.7.1

pdf2image==1.16.3
nltk==3.9.1
sentence-transformers==3.2.1